# More efficient decoding of string values