Tested with Scala 3 and JDK 21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                             (size)   Mode  Cnt        Score       Error  Units
StringOfAsciiCharsReading.zioJson        128  thrpt    5  2564276.761 ± 12171.566  ops/s
StringOfEscapedCharsReading.zioJson      128  thrpt    5   772746.093 ± 56346.729  ops/s
StringOfNonAsciiCharsReading.zioJson     128  thrpt    5  1534866.952 ± 14050.482  ops/s
```

#### After
```txt
Benchmark                             (size)   Mode  Cnt        Score       Error  Units
StringOfAsciiCharsReading.zioJson        128  thrpt    5  3804646.804 ±  7068.611  ops/s
StringOfEscapedCharsReading.zioJson      128  thrpt    5   794500.421 ± 24975.739  ops/s
StringOfNonAsciiCharsReading.zioJson     128  thrpt    5  1842774.078 ± 10740.569  ops/s
```